+++ indentation/src/Text/Parsec/Indentation/Char.hs
@@ -1,0 +1,4 @@
+{-# language PackageImports #-}
+module Text.Parsec.Indentation.Char (module Impl) where
+
+import "indentation-parsec" Text.Parsec.Indentation.Char as Impl