+++ indentation/src/Text/Parsec/Indentation/Token.hs
@@ -1,0 +1,4 @@
+{-# language PackageImports #-}
+module Text.Parsec.Indentation.Token (module Impl) where
+
+import "indentation-parsec" Text.Parsec.Indentation.Token as Impl