+++ indentation/src/Text/Parser/Indentation/Implementation.hs
@@ -1,0 +1,6 @@
+{-# language PackageImports #-}
+
+module Text.Parser.Indentation.Implementation (module Impl)
+       where
+
+import "indentation-core" Text.Parser.Indentation.Implementation as Impl